move components to seperate file and edit color
