+++ src/App.css
@@ -37,65 +37,3 @@
   }
 }
 
-body {
-  font: 14px "Century Gothic", Futura, sans-serif;
-  margin: 20px;
-  text-align: center;
-}
-
-ol, ul {
-  padding-left: 30px;
-}
-
-.board-row:after {
-  clear: both;
-  content: "";
-  display: table;
-}
-
-.restart {
-  margin-top: 10px;
-}
-
-.status {
-  margin-bottom: 10px;
-}
-
-.square {
-  background: #fff;
-  border: 1px solid #999;
-  float: left;
-  font-size: 24px;
-  font-weight: bold;
-  line-height: 34px;
-  height: 100px;
-  margin-right: -1px;
-  margin-top: -1px;
-  padding: 0;
-  text-align: center;
-  width: 100px;
-}
-
-.square:focus {
-  outline: none;
-}
-
-.kbd-navigation .square:focus {
-  background: #ddd;
-}
-
-.game {
-  display: flex;
-  flex-direction: row;
-}
-
-.game-info {
-  margin-left: 20px;
-}
-
-.board {
-  display: grid;
-  grid-template-columns: 100px 100px 100px;
-  grid-template-rows: auto; 
-  justify-content: center;
-}
