+++ src/components/Board.css
@@ -1,0 +1,72 @@
+body {
+    font: 14px "Century Gothic", Futura, sans-serif;
+    margin: 20px;
+    text-align: center;
+    background-color: #0d385e;
+    color: #fff;
+  }
+  
+  ol, ul {
+    padding-left: 30px;
+  }
+  
+  .board-row:after {
+    clear: both;
+    content: "";
+    display: table;
+  }
+  
+  .restart {
+    margin-top: 10px;
+    font-size: 25px;
+    border-radius: 4px;
+    padding: 14px 40px;
+    width: 300px;
+    color: #fff;
+    background-color: #ea2d9b;
+    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2), 0 6px 20px 0 rgba(0,0,0,0.19);
+  }
+  
+  .status {
+    margin-bottom: 10px;
+    font-size: 20px;
+  }
+  
+  .square {
+    background: rgb(114, 155, 221);
+    border: 1px solid rgb(0, 0, 0);
+    float: left;
+    font-size: 24px;
+    font-weight: bold;
+    line-height: 34px;
+    height: 100px;
+    margin-right: -1px;
+    margin-top: -1px;
+    padding: 0;
+    text-align: center;
+    width: 100px;
+  }
+  
+  .square:focus {
+    outline: none;
+  }
+  
+  .kbd-navigation .square:focus {
+    background: #ddd;
+  }
+  
+  .game {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .game-info {
+    margin-left: 20px;
+  }
+  
+  .board {
+    display: grid;
+    grid-template-columns: repeat(3, 100px);
+    grid-template-rows: auto; 
+    justify-content: center;
+  }
