+++ src/components/Square.js
@@ -1,0 +1,8 @@
+function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+export default Square;
